+++ frontend/src/app/(app)/(calendar)/calendar/page.tsx
@@ -1,6 +1,6 @@
 import { CardBodyWithLeftSidebar } from "~/features/appLayout/CardBodyWithLeftSidebar";
 import PrivateScheduleDayTimeline from "~/features/dayTimeline/PrivateScheduleDayTimeline";
-import { ViewerLeftSideBar } from "~/features/leftSidebar2/ViewerLeftSidebar";
+import { ViewerLeftSideBar } from "~/features/leftSidebar/ViewerLeftSidebar";
 
 export default function Page() {
   return (
