+++ frontend/src/app/practice/test4/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { ViewerLeftSideBar } from "~/features/leftSidebar2/ViewerLeftSidebar";
+import { ViewerLeftSideBar } from "~/features/leftSidebar/ViewerLeftSidebar";
 
 export default function Page() {
   return (
